+++ examples/flag_and_option_specifications.py
@@ -11,9 +11,9 @@ import sys
 
 # constants
 
-version = [ 0, 0, 1 ]
+VERSION = [ 0, 0, 2 ]
 
-info_lines = (
+INFO_LINES = (
 
     'CLASP.Python examples',
     ':version:',
@@ -39,31 +39,31 @@ specifications = (
 
 args = clasp.parse(sys.argv, specifications)
 
-if args.flagIsSpecified('--help'):
+if args.flag_is_specified('--help'):
 
-    clasp.show_usage(specifications, exit_code=0, version=version, stream=sys.stdout, info_lines = info_lines)
+    clasp.show_usage(specifications, exit_code=0, version=VERSION, stream=sys.stdout, info_lines=INFO_LINES)
 
-if args.flagIsSpecified('--version'):
+if args.flag_is_specified('--version'):
 
-    clasp.show_version(specifications, exit_code=0, version=version, stream=sys.stdout)
+    clasp.show_version(specifications, exit_code=0, version=VERSION, stream=sys.stdout)
 
 
 # Program-specific processing of flags/options
 
-opt = args.lookupOption('--verbosity')
+opt = args.lookup_option('--verbosity')
 if (opt):
 
     sys.stdout.write("verbosity is specified as: %s\n" % opt.value)
 
 
-if (args.flagIsSpecified('--debug')):
+if (args.flag_is_specified('--debug')):
 
     sys.stdout.write("Debug mode is specified\n")
 
 
 # Check for any unrecognised flags or options
 
-unused = args.getFirstUnusedFlagOrOption();
+unused = args.get_first_unused_flag_or_option();
 if (unused):
 
     sys.stderr.write("%s: unrecognised flag/option: %s\n" % (args.program_name, unused))